+++ samplesheets/tests/test_commands.py
@@ -18,8 +18,18 @@ from projectroles.tests.test_models import (
 # Sodarcache dependency
 from sodarcache.models import JSONCacheItem
 
-from samplesheets.models import GenericMaterial
-from samplesheets.rendering import STUDY_TABLE_CACHE_ITEM
+# Timeline dependency
+from timeline.models import ProjectEvent
+
+from samplesheets.management.commands.normalizesheets import (
+    LIB_NAME,
+    LIB_NAME_REPLACE,
+)
+from samplesheets.models import GenericMaterial, ISATab
+from samplesheets.rendering import (
+    SampleSheetTableBuilder,
+    STUDY_TABLE_CACHE_ITEM,
+)
 from samplesheets.tests.test_io import (
     SampleSheetIOMixin,
     SHEET_DIR,
@@ -36,7 +46,135 @@ SHEET_PATH_ALT = SHEET_DIR + 'i_small2.zip'
 ALT_NAMES_INVALID = ['XXX', 'YYY', 'ZZZ']
 
 
-class TestSyncnamesCommand(
+class TestNormalizesheets(
+    ProjectMixin, RoleMixin, RoleAssignmentMixin, SampleSheetIOMixin, TestCase
+):
+    """Tests for the normalizesheets command"""
+
+    def _assert_material_header(self, materials, header, expected):
+        """Assert count of materials which contain a specific header name"""
+        self.assertEqual(
+            materials.filter(headers__icontains=header).count(),
+            expected,
+        )
+
+    def _assert_study_table_header(self, study_tables, assay, header, expected):
+        """
+        Assert count of assay table headers which contain a specific header
+        name.
+        """
+        a = str(assay.sodar_uuid)
+        h_name = header.lower()
+        top_header = study_tables['assays'][a]['top_header']
+        self.assertEqual(
+            len(
+                [
+                    h
+                    for h in top_header
+                    if h['value'].lower() == h_name
+                    and h['headers'][0].lower() == h_name
+                ]
+            ),
+            expected,
+        )
+
+    def _assert_tl_event(self, expected):
+        self.assertEqual(
+            ProjectEvent.objects.filter(
+                event_name='sheet_normalize', project=self.project
+            ).count(),
+            expected,
+        )
+
+    def setUp(self):
+        # Init roles
+        self.init_roles()
+        # Make owner user
+        self.user_owner = self.make_user('owner')
+        # Init project and assignment
+        self.project = self.make_project(
+            'TestProject', SODAR_CONSTANTS['PROJECT_TYPE_PROJECT'], None
+        )
+        self.owner_as = self.make_assignment(
+            self.project, self.user_owner, self.role_owner
+        )
+        # Import investigation
+        self.investigation = self.import_isa_from_file(SHEET_PATH, self.project)
+        self.study = self.investigation.studies.first()
+        self.assay = self.study.assays.first()
+        # Set up study tables in cache
+        self.tb = SampleSheetTableBuilder()
+        self.cache_backend = get_backend_api('sodar_cache')
+        self.cache_name = STUDY_TABLE_CACHE_ITEM.format(
+            study=self.study.sodar_uuid
+        )
+        self.cache_args = [APP_NAME, self.cache_name, self.project]
+        self.tb.get_study_tables(self.study)
+
+    def test_command(self):
+        """Test normalizesheets"""
+        # Materials
+        materials = GenericMaterial.objects.filter(assay=self.assay)
+        self._assert_material_header(materials, LIB_NAME, 2)
+        self._assert_material_header(materials, LIB_NAME_REPLACE, 0)
+        # Cached study tables
+        cache_item = self.cache_backend.get_cache_item(*self.cache_args)
+        self._assert_study_table_header(
+            cache_item.data, self.assay, LIB_NAME, 1
+        )
+        self._assert_study_table_header(
+            cache_item.data, self.assay, LIB_NAME_REPLACE, 0
+        )
+        # Sheet version
+        self.assertEqual(ISATab.objects.count(), 1)
+        # Timeline event
+        self._assert_tl_event(0)
+        call_command('normalizesheets')
+        materials = GenericMaterial.objects.filter(assay=self.assay)
+        self._assert_material_header(materials, LIB_NAME, 0)
+        self._assert_material_header(materials, LIB_NAME_REPLACE, 2)
+        cache_item = self.cache_backend.get_cache_item(*self.cache_args)
+        self._assert_study_table_header(
+            cache_item.data, self.assay, LIB_NAME, 0
+        )
+        self._assert_study_table_header(
+            cache_item.data, self.assay, LIB_NAME_REPLACE, 1
+        )
+        # Sheet version
+        self.assertEqual(ISATab.objects.count(), 2)
+        # Timeline event
+        self._assert_tl_event(1)
+
+    def test_command_check(self):
+        """Test normalizesheets with check mode"""
+        materials = GenericMaterial.objects.filter(assay=self.assay)
+        self._assert_material_header(materials, LIB_NAME, 2)
+        self._assert_material_header(materials, LIB_NAME_REPLACE, 0)
+        cache_item = self.cache_backend.get_cache_item(*self.cache_args)
+        self._assert_study_table_header(
+            cache_item.data, self.assay, LIB_NAME, 1
+        )
+        self._assert_study_table_header(
+            cache_item.data, self.assay, LIB_NAME_REPLACE, 0
+        )
+        self.assertEqual(ISATab.objects.count(), 1)
+        self._assert_tl_event(0)
+        call_command('normalizesheets', check=True)
+        materials = GenericMaterial.objects.filter(assay=self.assay)
+        self._assert_material_header(materials, LIB_NAME, 2)
+        self._assert_material_header(materials, LIB_NAME_REPLACE, 0)
+        cache_item = self.cache_backend.get_cache_item(*self.cache_args)
+        self._assert_study_table_header(
+            cache_item.data, self.assay, LIB_NAME, 1
+        )
+        self._assert_study_table_header(
+            cache_item.data, self.assay, LIB_NAME_REPLACE, 0
+        )
+        self.assertEqual(ISATab.objects.count(), 1)
+        self._assert_tl_event(0)
+
+
+class TestSyncnames(
     ProjectMixin, RoleMixin, RoleAssignmentMixin, SampleSheetIOMixin, TestCase
 ):
     """Tests for the syncnames command"""
@@ -70,7 +208,7 @@ class TestSyncnamesCommand(
             self.assertEqual(m.alt_names, get_alt_names(m.name))
 
 
-class TestSyncstudytablesCommand(
+class TestSyncstudytables(
     ProjectMixin, RoleMixin, RoleAssignmentMixin, SampleSheetIOMixin, TestCase
 ):
     """Tests for the syncstudytables command"""